pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-recruiter"
version = "0.1.0"
description = "AI-Powered Multi-Agent Recruitment System using CrewAI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Saikrishna Paila", email = "saikrishnapaila15@gmail.com"}
]
keywords = [
    "ai",
    "recruitment",
    "multi-agent",
    "crewai",
    "automation",
    "hiring",
    "resume-screening",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"

dependencies = [
    # CrewAI - Multi-agent framework
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",

    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "groq>=0.11.0",

    # Document Processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "unstructured>=0.15.0",

    # Embeddings & ML
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",

    # Vector Database
    "chromadb>=0.5.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",

    # Configuration
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # Email (optional)
    "resend>=2.0.0",

    # Utilities
    "rich>=13.9.0",
    "typer>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

calendar = [
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]

all = [
    "smart-recruiter[dev]",
    "smart-recruiter[calendar]",
]

[project.scripts]
smart-recruiter = "smart_recruiter.cli:app"
recruiter = "smart_recruiter.cli:app"

[project.urls]
Homepage = "https://github.com/saikrishnapaila/smart-recruiter-agent"
Documentation = "https://github.com/saikrishnapaila/smart-recruiter-agent#readme"
Repository = "https://github.com/saikrishnapaila/smart-recruiter-agent"
Issues = "https://github.com/saikrishnapaila/smart-recruiter-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["smart_recruiter*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
